La Salle Paper-Cut Printer Queue
Used to bring the printer queue for the Paper-Cut printer service that La Salle uses for all of its printers on campus. Made it quicker to open the printer queue, especially when set to start at sign-on with windows event scheduler, to see if there were issues when students, staff, faculty, etc. try to print.