Look-up La Salle User Info
Used to look up a user through the windows command line via the La Salle domain. Created to look-up information about the user that we did not within our library record of them in FOLIO. Gave us the ability to get information such as: If their account was active, if they needed to reset their password, if they had access to computers on campus, their last successful logon, and what group memberships they belong to.