All La Salle users on domain
Used to generate a .txt file of all users that are on the La Salle domain. This was helpful to get a more real idea of if a user was part of La Salle University. We have a few methods to check for a user's record, but each of them where not in real time. Our own library database via FOLIO that was used was updated in batches, every 1-3 weeks, and would miss if a student/staff was added late. La Salle's online directory, which also was close, but not real time. This file would check the domain in real time to return all users to provide an accurate list of users at La Salle